Mostly HomingCycle
Added more up-to-date variables and definition into the homing cycle code. Changed homing cycle to use the provided SpeedyStepper code. The case will cycle back to the start of loop() after each motor is homed. If any of the homing functions fail, they will return a false which is used to call an error.

Co-Authored-By: linhngo11 <8979507+linhngo11@users.noreply.github.com>