statemachine.c initial setup and some HomingCycle changes
The bulk of this update is for statemachine. I added all of the starting code, setup, and loop so that we have a stable reference to use when declaring values as we write cases. I copied most everything from Robert's code but ENSURE THIS GETS A LOOK OVER AND APPROVAL BY ROBERT BEFORE DEPENDING ON IT.

The HomingCycle changes are minimal - just some spelling corrections and a change to one variable.

Co-Authored-By: Robwgit <79285262+Robwgit@users.noreply.github.com>