Update HomingCycle, add to statemachine, and check for misstep
I created conditions for misstep errors. I have some questions about Robert's movement code that are necessary to figure out where I should out in the misstep flag. There is also a question of whether an error condition should immediately be called if a misstep happens. HomingCycle was updated to use the misstep flag for stream-lining and then added to statemachine.