Part 2 of major code review
A lot of minor changes to when and which LEDs are called. I'm waiting on coorspondence with Robert to figure out what to do with the Camera cases but I am confident they won't work in their current state.  Also added WIS TRAY_MOVE_IN cases to make sure the tray is in the out position before WAIT_TO_START is called.