Correction to print statements
Greg suggested that the repeat Serial.print statements are probably slowing down the actuator movement and may be the cause of the program getting stuck. I changed the print statements so that they only occur once whenever the currState value is changed. More testing required to see how this affects the Arduino operation.